compress aggressively
